requirements.txt
aws-cdk-lib==2.233.0
constructs==10.4.4
cdk-watchful>=0.6.233
pytest>=7.0.0
